Refactor: 멤버 컨트롤러에 존재하던 엔티티 서비스로 리펙토링
mapper 클래스 추가 및 테스트 코드 수정
